Moved Filters to their own namespace since there are a few common ones
