+++ src/Application/Filter.php
@@ -8,33 +8,12 @@
 
 namespace Nexmo\Application;
 
-use Nexmo\Entity\FilterInterface;
+use Nexmo\Entity\Filter\DateFilter;
 
 /**
  * Simple value object for application filtering.
+ * @deprecated Please use Nexmo\Entity\Filter\DateFilter instead
  */
-class Filter implements FilterInterface
+class Filter extends DateFilter
 {
-    const FORMAT = 'Y:m:d:H:i:s';
-
-    protected $start;
-    protected $end;
-
-    public function __construct(\DateTime $start, \DateTime $end)
-    {
-        if ($start < $end) {
-            $this->start = $start;
-            $this->end = $end;
-        } else {
-            $this->start = $end;
-            $this->end = $start;
-        }
-    }
-
-    public function getQuery()
-    {
-        return [
-            'date' => $this->start->format(self::FORMAT) . '-' . $this->end->format(self::FORMAT)
-        ];
-    }
 }
